Revert "Feature create context"
