+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.1.1",
-    "date-fns": "^2.30.0",
     "phosphor-react": "^1.4.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
