+++ src/App.tsx
@@ -3,15 +3,12 @@ import { defaultTheme } from './styles/themes/default'
 import { GlobalStyle } from './styles/global'
 import { Router } from './Router'
 import { BrowserRouter } from 'react-router-dom'
-import { CyclesContextProvider } from './contexts/CyclesContext'
 
 export function App() {
   return (
     <ThemeProvider theme={defaultTheme}>
       <BrowserRouter>
-        <CyclesContextProvider>
-          <Router />
-        </CyclesContextProvider>
+        <Router />
       </BrowserRouter>
       <GlobalStyle />
     </ThemeProvider>
