+++ src/Pages/History/index.tsx
@@ -1,16 +1,9 @@
-import { useContext } from 'react'
 import { HistoryContainer, HistoryList, Status } from './styles'
-import { CyclesContext } from '../../contexts/CyclesContext'
 
 export function History() {
-  const { cycles } = useContext(CyclesContext)
-
   return (
     <HistoryContainer>
       <h1>Meu Histórico</h1>
-
-      <pre>{JSON.stringify(cycles, null, 2)}</pre>
-
       <HistoryList>
         <table>
           <thead>
