--- src/contexts/CyclesContext.tsx
@@ -1,104 +1,0 @@
-import React, { createContext, useState } from 'react'
-
-interface CreateCycleData {
-  task: string
-  minutesAmount: number
-}
-
-interface Cycle {
-  id: string
-  task: string
-  minutesAmount: number
-  startDate: Date
-  interruptedDate?: Date
-  finisheDate?: Date
-}
-
-interface CyclesContextType {
-  cycles: Cycle[]
-  activeCycle: Cycle | undefined
-  activeCycleId: string | null
-  amountSecondsPassed: number
-  markCurrentCyclesAsFinished: () => void
-  serSecondsPassed: (seconds: number) => void
-  createNewCycle: (data: CreateCycleData) => void
-  interruptCurrentCycle: () => void
-}
-
-export const CyclesContext = createContext({} as CyclesContextType)
-
-interface CyclesContextProviderProps {
-  children: React.ReactNode
-}
-
-export function CyclesContextProvider({
-  children,
-}: CyclesContextProviderProps) {
-  const [cycles, setCycles] = useState<Cycle[]>([])
-  const [activeCycleId, setActiveCycleId] = useState<string | null>(null)
-  const [amountSecondsPassed, setAmountSecondsPassed] = useState(0)
-
-  const activeCycle = cycles.find((cycle) => cycle.id === activeCycleId)
-
-  function serSecondsPassed(seconds: number) {
-    setAmountSecondsPassed(seconds)
-  }
-
-  function markCurrentCyclesAsFinished() {
-    setCycles((state) =>
-      state.map((cycle) => {
-        if (cycle.id === activeCycleId) {
-          return { ...cycle, finishedDate: new Date() }
-        } else {
-          return cycle
-        }
-      }),
-    )
-  }
-
-  function createNewCycle(data: CreateCycleData) {
-    const id = String(new Date().getTime())
-
-    const newCycle: Cycle = {
-      id,
-      task: data.task,
-      minutesAmount: data.minutesAmount,
-      startDate: new Date(),
-    }
-
-    setCycles((state) => [...state, newCycle])
-    setActiveCycleId(id)
-    setAmountSecondsPassed(0)
-  }
-
-  function interruptCurrentCycle() {
-    setCycles((state) =>
-      state.map((cycle) => {
-        if (cycle.id === activeCycleId) {
-          return { ...cycle, interruptedDate: new Date() }
-        } else {
-          return cycle
-        }
-      }),
-    )
-
-    setActiveCycleId(null)
-  }
-
-  return (
-    <CyclesContext.Provider
-      value={{
-        cycles,
-        activeCycle,
-        activeCycleId,
-        amountSecondsPassed,
-        markCurrentCyclesAsFinished,
-        serSecondsPassed,
-        createNewCycle,
-        interruptCurrentCycle,
-      }}
-    >
-      {children}
-    </CyclesContext.Provider>
-  )
-}
